requirements.txt
streamlit==1.32.0
anthropic==0.39.0
python-docx==1.1.0
PyMuPDF==1.25.1
Pillow==10.2.0
python-dotenv==1.0.1
reportlab==4.1.0
httpx==0.27.0